페이지 클래스 추상?화?
