+++ src/EnumCollect/wiseSayingEnum.java
@@ -3,6 +3,7 @@ package EnumCollect;
 public enum wiseSayingEnum {
     KEYWORD("keyword"),
     KEYWORD_TYPE("keywordType"),
+    PAGE("page"),
     ID("id");
     private final String query;
     wiseSayingEnum(String query){
